1:完成订单得预约服务

+++ src/main/java/com/haige/db/entity/GoodsInfoDO.java
@@ -20,7 +20,7 @@ public class GoodsInfoDO {
 
     private String goodsTitle;
 
-    private Byte goodsFreq;
+    private Integer goodsFreq;
 
     private Date goodsExpreDate;
